v1.06 - The End (#16)
This is the end.

Created a new ending sequence, so Maui gets a proper closure if he dies.
Created a new ending sequence, so Maui gets a proper victory if he wins.
Added Water Color fluctuations around the Star of Tangaroa to highlight the location more.
Added Grandma Dialogue
She now gives you the Rod of Legends which gives you a 100% fishing success rate as well as open up the ending.
Fixed up error messages
Most of them should now be Color.write, with ERROR highlighting and 2 spaces afterwards.
Crashing into terrain now says "You cant move" instead of sail, as this didn't work in the cave level very well.
~ Changed a lot of functions to include the Item list now.
~ Fixed a bug where the game would crash in tutorial level
This is likely the final update for a long while, at least until I decide to open up the game again. Though this journey wasn't as long and as perilous as Maui's, it was still a journey. With this update adding a proper conclusion, it seems fitting that this would be the final update.
To whomever is reading this right now, Thank you for being on this journey with me. Ka kite anō au i a koe.